backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
psycopg2-binary==2.9.9  # For PostgreSQL (Render default)
python-decouple==3.8
dj-database-url==2.1.0  # For parsing DATABASE_URL
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0

# PDF Generation
weasyprint==60.2
reportlab==4.0.7

# Email
django-ses==3.5.0
sendgrid==6.11.0

# SMS & WhatsApp
twilio==8.10.0
requests==2.31.0

# Payment Gateways
razorpay==1.4.1
stripe==7.0.0

# Calendar Integration
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# File Storage
boto3==1.29.7  # AWS S3
google-cloud-storage==2.14.0

# Maps
googlemaps==4.10.0

# Utilities
celery==5.3.4  # For background tasks
redis==5.0.1  # For Celery broker
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0

